Updated s.t. gui works and cli works
