+++ src/org/traxgame/gui/LoadingFrame.java
@@ -1,0 +1,24 @@
+package org.traxgame.gui;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class LoadingFrame extends JFrame {
+	/**
+	 *
+	 */
+	private static final long serialVersionUID = 2221062376560774087L;
+
+	public LoadingFrame(JFrame owner) {
+		//super(owner);
+		this.add(new Label("AI is thinking. Please wait...."));
+		this.pack();
+		int x = owner.getX() + (owner.getWidth() - getPreferredSize().width)
+				/ 2;
+		int y = owner.getY() + (owner.getHeight() - getPreferredSize().height)
+				/ 2;
+		setLocation(x, y);
+
+		repaint();
+	}
+}
